renaming `query_result_struct.h`
