+++ src/VecSim/query_result_definitions.h
