+++ tests/unit/test_hnsw_parallel.cpp
@@ -8,7 +8,7 @@
 #include "VecSim/vec_sim.h"
 #include "VecSim/algorithms/hnsw/hnsw_single.h"
 #include "test_utils.h"
-#include "VecSim/query_result_struct.h"
+#include "VecSim/query_result_definitions.h"
 #include <unistd.h>
 #include <random>
 #include <thread>
